requirements.txt

Flask
numpy
opencv-python
pillow
deepface
werkzeug
gunicorn

